Feat(InputForm): Input 컴포넌트 ui, 로직 분리
